Optimization, cleanup, QA

+++ Inpsyde/Sniffs/CodeQuality/ElementNameMinimalLengthSniff.php
@@ -1,19 +1,5 @@
 <?php
 
-/*
- * This file is part of the php-coding-standards package.
- *
- * (c) Inpsyde GmbH
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- *
- * This file contains code from "phpcs-calisthenics-rules" repository
- * found at https://github.com/object-calisthenics
- * Copyright (c) 2014 Doctrine Project
- * released under MIT license.
- */
-
 declare(strict_types=1);
 
 namespace Inpsyde\Sniffs\CodeQuality;
@@ -22,10 +8,6 @@ use Inpsyde\PhpcsHelpers;
 use PHP_CodeSniffer\Files\File;
 use PHP_CodeSniffer\Sniffs\Sniff;
 
-/**
- * @package php-coding-standards
- * @license http://opensource.org/licenses/MIT MIT
- */
 class ElementNameMinimalLengthSniff implements Sniff
 {
     /**
@@ -69,10 +51,16 @@ class ElementNameMinimalLengthSniff implements Sniff
     public $additionalAllowedNames = [];
 
     /**
-     * @return int[]
+     * @return array<int>
+     *
+     * phpcs:disable Inpsyde.CodeQuality.ArgumentTypeDeclaration
+     * phpcs:disable Inpsyde.CodeQuality.ReturnTypeDeclaration
      */
     public function register()
     {
+        // phpcs:enable Inpsyde.CodeQuality.ArgumentTypeDeclaration
+        // phpcs:enable Inpsyde.CodeQuality.ReturnTypeDeclaration
+
         return [T_CLASS, T_TRAIT, T_INTERFACE, T_CONST, T_FUNCTION, T_VARIABLE];
     }
 
@@ -80,11 +68,17 @@ class ElementNameMinimalLengthSniff implements Sniff
      * @param File $file
      * @param int $position
      * @return void
+     *
+     * phpcs:disable Inpsyde.CodeQuality.ArgumentTypeDeclaration
+     * phpcs:disable Inpsyde.CodeQuality.ReturnTypeDeclaration
      */
     public function process(File $file, $position)
     {
+        // phpcs:enable Inpsyde.CodeQuality.ArgumentTypeDeclaration
+        // phpcs:enable Inpsyde.CodeQuality.ReturnTypeDeclaration
+
         $elementName = PhpcsHelpers::tokenName($file, $position);
-        $elementNameLength = mb_strlen($elementName);
+        $elementNameLength = (int)mb_strlen($elementName);
 
         if ($this->shouldBeSkipped($elementNameLength, $elementName)) {
             return;
@@ -102,11 +96,20 @@ class ElementNameMinimalLengthSniff implements Sniff
         $file->addError($message, $position, 'TooShort');
     }
 
+    /**
+     * @param int $elementNameLength
+     * @param string $elementName
+     * @return bool
+     */
     private function shouldBeSkipped(int $elementNameLength, string $elementName): bool
     {
         return ($elementNameLength >= $this->minLength) || $this->isShortNameAllowed($elementName);
     }
 
+    /**
+     * @param string $variableName
+     * @return bool
+     */
     private function isShortNameAllowed(string $variableName): bool
     {
         $target = strtolower($variableName);
